Add Pollo recipe page, update home, add local img
Add new Pollo Asado recipe page in the same style as the Carne Asada
Adobada page.  Update the homepage by adding a link to the Pollo Asado
page while also putting the existing Carne Asada Adobada link in an
unordered list with it.  Save a cropped picture from an online source
locally in created images folder to be referenced in the Pollo Asado
page.